+++ LuckyoneUI/Modules/Skins/Addons/NovaSpellRankChecker.lua
@@ -19,8 +19,6 @@ function Private:Skin_NovaSpellRankChecker()
 		-- Skin and resize the spell rank checker button
 		S:HandleButton(SpellBookFrameButton)
 		SpellBookFrameButton:SetFrameStrata('HIGH')
-		SpellBookFrameButton:Width(120)
-		SpellBookFrameButton:Height(26)
 
 		-- Move it to the bottom left of the spell book frame
 		SpellBookFrameButton:ClearAllPoints()